requirements.txt
langchain_core==0.3.23
langchain_openai==0.2.12
pydantic==2.10.3
python-dotenv==1.0.1
